Styled components into seperate component files.
