+++ src/components/CountryPicker/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "CountryPicker.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./CountryPicker.js"
+}
