+++ src/components/Emoji/Emoji.js
@@ -1,0 +1,10 @@
+// @flow
+import styled from 'styled-components/native';
+
+const Emoji = styled.Text`
+  flex: 0 0 40px;
+  font-size: 36px;
+  line-height: 50px;
+`;
+
+export default Emoji;
