+++ src/components/Emoji/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Emoji.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./Emoji.js"
+}
