+++ src/components/FooterText/FooterText.js
@@ -1,0 +1,10 @@
+// @flow
+import styled from 'styled-components/native';
+import { Label } from 'components/Typography';
+
+const FooterText = styled(Label)`
+  text-align: center;
+  max-width: 300px;
+`;
+
+export default FooterText;
