+++ src/components/FooterText/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "FooterText.js",
+  "version": "0.0.0",
+  "main": "./FooterText.js"
+}
