+++ src/components/Form/Form.js
@@ -1,0 +1,20 @@
+// @flow
+import styled from 'styled-components/native';
+import { Form, Input } from 'native-base';
+
+export const LoginForm = styled(Form)`
+  margin: 10px 0 40px;
+`;
+
+export const InputField = styled(Input)`
+  border-bottom-width: 1px;
+  border-color: rgb(151,151,151);
+  font-size: 24px;
+  margin-bottom: 10px;
+`;
+
+export const PhoneInput = styled(Input)`
+  border-bottom-width: 1px;
+  border-color: rgb(151,151,151);
+  font-size: 24px;
+`;
