+++ src/components/Form/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "Form.js",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./Form.js"
+}
