+++ src/screens/Signup/Signup.js
@@ -2,13 +2,16 @@
 import * as React from 'react';
 import type { NavigationScreenProp } from 'react-navigation';
 import { OTP_STATUS } from 'constants/navigationConstants';
-import styled from 'styled-components/native';
 import { Container, Footer, Center } from 'components/Layout';
 import Wrapper from 'components/Wrapper';
 import Button from 'components/Button';
+import Emoji from 'components/Emoji';
 import HyperLink from 'components/HyperLink';
+import FooterText from 'components/FooterText';
+import { CountryPicker, CountryPickerWrapper } from 'components/CountryPicker';
 import { Title, Body, Label } from 'components/Typography';
-import { Form, Picker, Icon, Input } from 'native-base';
+import { LoginForm, PhoneInput, InputField } from 'components/Form';
+import { Picker, Icon } from 'native-base';
 import countries from 'utils/countries.json';
 
 type Props = {
@@ -20,49 +23,6 @@ type State = {
   selectedCountryFlag: string,
   selectedCountryCallingCode: string,
 }
-
-const LoginForm = styled(Form)`
-  margin: 10px 0 40px;
-`;
-
-const InputField = styled(Input)`
-  border-bottom-width: 1px;
-  border-color: rgb(151,151,151);
-  font-size: 24px;
-  margin-bottom: 10px;
-
-`;
-
-const PhoneInput = styled(Input)`
-  border-bottom-width: 1px;
-  border-color: rgb(151,151,151);
-  font-size: 24px;
-`;
-
-const Emoji = styled.Text`
-  flex: 0 0 40px;
-  font-size: 36px;
-  line-height: 50px;
-`;
-
-const CountryPicker = styled(Picker)`
-  flex: 1;
-  align-self: flex-end;
-`;
-
-const CountryPickerWrapper = styled.View`
-  width: 100%;
-  margin-bottom: 10px;
-  flex-direction: row;
-  justify-content: space-between;
-  border-bottom-width: 1px;
-  border-color: rgb(151,151,151);
-`;
-
-const FooterText = styled(Label)`
-  text-align: center;
-  max-width: 300px;
-`;
 
 class Signup extends React.Component<Props, State> {
   state = {
